+++ example/landformSiteOverlay.js
@@ -1,7 +1,7 @@
 import {
 	TilesRenderer,
 	EnvironmentControls,
-} from '../src/index.js';
+} from '3d-tiles-renderer';
 import {
 	Scene,
 	WebGLRenderer,