+++ example/r3f/globe.jsx
@@ -2,10 +2,17 @@ import { StrictMode, useRef } from 'react';
 import { createRoot } from 'react-dom/client';
 
 // TilesRenderer, controls and attribution imports
-import { TilesPlugin, TilesRenderer, TilesAttributionOverlay, GlobeControls, EastNorthUpFrame } from '../../src/r3f/index.jsx';
+import {
+	TilesPlugin,
+	TilesRenderer,
+	TilesAttributionOverlay,
+	GlobeControls,
+	EastNorthUpFrame,
+	CompassGizmo,
+} from '3d-tiles-renderer/r3f';
 
 // Plugins
-import { GoogleCloudAuthPlugin } from '../../src/index.js';
+import { GoogleCloudAuthPlugin } from '3d-tiles-renderer';
 import { GLTFExtensionsPlugin } from '../src/plugins/GLTFExtensionsPlugin.js';
 import { TilesFadePlugin } from '../src/plugins/fade/TilesFadePlugin.js';
 import { TileCompressionPlugin } from '../src/plugins/TileCompressionPlugin.js';
@@ -14,7 +21,7 @@ import { DRACOLoader } from 'three/examples/jsm/loaders/DRACOLoader.js';
 
 // R3F, DREI and LEVA imports
 import { Canvas, useFrame } from '@react-three/fiber';
-import { Environment, GizmoHelper, GizmoViewport } from '@react-three/drei';
+import { Environment } from '@react-three/drei';
 import { useControls } from 'leva';
 import { MathUtils, Vector3 } from 'three';
 
@@ -64,7 +71,7 @@ function App() {
 	return (
 		<Canvas
 			camera={ {
-				position: [ 0, 4 * 1e7, 1e8 ],
+				position: [ 0, 0.5 * 1e7, 1.5 * 1e7 ],
 			} }
 			style={ {
 				width: '100%',
@@ -74,7 +81,7 @@ function App() {
 				left: 0,
 				top: 0,
 			} }
-			color='red'
+			flat
 		>
 			<color attach="background" args={ [ 0x111111 ] } />
 
@@ -82,27 +89,27 @@ function App() {
 				3D Tiles renderer tile set
 				Use a "key" property to ensure the tiles renderer gets recreated when the api token or asset change
 			*/}
-			<group rotation-x={ - Math.PI / 2 }>
-				<TilesRenderer key={ apiToken }>
-					<TilesPlugin plugin={ GoogleCloudAuthPlugin } args={ { apiToken } } />
-					<TilesPlugin plugin={ GLTFExtensionsPlugin } dracoLoader={ dracoLoader } />
-					<TilesPlugin plugin={ TileCompressionPlugin } />
-					<TilesPlugin plugin={ UpdateOnChangePlugin } />
-					<TilesPlugin plugin={ TilesFadePlugin } />
+			<TilesRenderer key={ apiToken } group={ { rotation: [ - Math.PI / 2, 0, 0 ] } }>
+				<TilesPlugin plugin={ GoogleCloudAuthPlugin } args={ { apiToken } } />
+				<TilesPlugin plugin={ GLTFExtensionsPlugin } dracoLoader={ dracoLoader } />
+				<TilesPlugin plugin={ TileCompressionPlugin } />
+				<TilesPlugin plugin={ UpdateOnChangePlugin } />
+				<TilesPlugin plugin={ TilesFadePlugin } />
 
-					{/* Controls */}
-					<GlobeControls enableDamping={ true } />
+				{/* Controls */}
+				<GlobeControls enableDamping={ true } />
 
-					{/* Attributions */}
-					<TilesAttributionOverlay />
+				{/* Attributions */}
+				<TilesAttributionOverlay />
 
-					{/* Pointer to NASA JPL */}
-					<EastNorthUpFrame lat={ 34.2013 * MathUtils.DEG2RAD } lon={ - 118.1714 * MathUtils.DEG2RAD } height={ 350 }>
-						<Pointer />
-					</EastNorthUpFrame>;
-				</TilesRenderer>
-			</group>
+				{/* Pointer to NASA JPL */}
+				<EastNorthUpFrame lat={ 34.2013 * MathUtils.DEG2RAD } lon={ - 118.1714 * MathUtils.DEG2RAD } height={ 350 }>
+					<Pointer />
+				</EastNorthUpFrame>;
 
+				{/* Add compass gizmo */}
+				<CompassGizmo />
+			</TilesRenderer>
 
 			{/* other r3f staging */}
 			<Environment
@@ -110,10 +117,6 @@ function App() {
 				backgroundBlurriness={ 0.9 }
 				environmentIntensity={ 1 }
 			/>
-			<GizmoHelper alignment="bottom-right">
-				<GizmoViewport />
-			</GizmoHelper>
-
 		</Canvas>
 	);