+++ example/r3f/ion.jsx
@@ -2,8 +2,8 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
 // TilesRenderer, controls and attribution imports
-import { TilesPlugin, TilesRenderer, TilesAttributionOverlay, EnvironmentControls } from '../../src/r3f/index.jsx';
-import { CesiumIonAuthPlugin } from '../../src/index.js';
+import { TilesPlugin, TilesRenderer, TilesAttributionOverlay, EnvironmentControls } from '3d-tiles-renderer/r3f';
+import { CesiumIonAuthPlugin } from '3d-tiles-renderer';
 
 // Plugins
 import { GLTFExtensionsPlugin } from '../src/plugins/GLTFExtensionsPlugin.js';