+++ src/base/traverseFunctions.js
@@ -340,6 +340,9 @@ export function markVisibleTiles( tile, renderer ) {
 	const emptyRootTile = tile.__depthFromRenderedParent === 0;
 	const allChildrenLoaded = tile.__allChildrenLoaded || emptyRootTile;
 
+	// TODO: decide if we should behave differently at root tiles
+	// const allChildrenLoaded = tile.__allChildrenLoaded;
+
 	// If we've met the SSE requirements and we can load content then fire a fetch.
 	const includeTile = meetsSSE || tile.refine === 'ADD';
 	if ( includeTile && ! loadedContent && ! lruCache.isFull() && hasContent ) {
@@ -372,7 +375,7 @@ export function markVisibleTiles( tile, renderer ) {
 
 	// If we're additive then don't stop the traversal here because it doesn't matter whether the children load in
 	// at the same rate.
-	if ( tile.refine === 'REPLACE' && meetsSSE && ! allChildrenLoaded && loadedContent ) {
+	if ( tile.refine === 'REPLACE' && meetsSSE && ! allChildrenLoaded ) {
 
 		// load the child content if we've found that we've been loaded so we can move down to the next tile
 		// layer when the data has loaded.