+++ src/r3f/utilities/useForceUpdate.jsx
@@ -1,0 +1,8 @@
+import { useState } from 'react';
+
+export function useForceUpdate() {
+
+	const [ value, setValue ] = useState( 0 );
+	return [ value, () => setValue( value + 1 ) ];
+
+}