+++ src/three/controls/GlobeControls.js
@@ -15,7 +15,6 @@ const _rotMatrix = new Matrix4();
 const _pos = new Vector3();
 const _vec = new Vector3();
 const _center = new Vector3();
-const _up = new Vector3();
 const _forward = new Vector3();
 const _right = new Vector3();
 const _targetRight = new Vector3();
@@ -28,9 +27,6 @@ const _ray = new Ray();
 const _ellipsoid = new Ellipsoid();
 
 const _pointer = new Vector2();
-const _prevPointer = new Vector2();
-const _deltaPointer = new Vector2();
-
 const MIN_ELEVATION = 400;
 
 export class GlobeControls extends EnvironmentControls {
@@ -57,11 +53,8 @@ export class GlobeControls extends EnvironmentControls {
 		this.useFallbackPlane = false;
 		this.reorientOnDrag = false;
 
-		this.inertiaAxis = new Vector3();
 		this.dragQuaternion = new Quaternion();
-		this.inertiaDragMode = 0;
 
-		this.allowNegativeNearPlanes = true;
 		this.setTilesRenderer( tilesRenderer );
 
 	}
@@ -171,11 +164,7 @@ export class GlobeControls extends EnvironmentControls {
 
 		}
 
-		const {
-			camera,
-			tilesGroup,
-			pivotMesh,
-		} = this;
+		const { camera, pivotMesh } = this;
 
 		// if we're outside the transition threshold then we toggle some reorientation behavior
 		// when adjusting the up frame while moving the camera
@@ -197,39 +186,21 @@ export class GlobeControls extends EnvironmentControls {
 		// fire basic controls update
 		super.update( deltaTime );
 
-		if ( camera.isPerspectiveCamera ) {
-
-			// clamp the camera distance
-			let distanceToCenter = this.getDistanceToCenter();
-			const maxDistance = this._getMaxPerspectiveDistance();
-			if ( distanceToCenter > maxDistance ) {
-
-				_vec.setFromMatrixPosition( tilesGroup.matrixWorld ).sub( camera.position ).normalize().multiplyScalar( - 1 );
-				camera.position.setFromMatrixPosition( tilesGroup.matrixWorld ).addScaledVector( _vec, maxDistance );
-				camera.updateMatrixWorld();
-
-				distanceToCenter = maxDistance;
-
-			}
-
-		}
-
 		// update the camera planes and the ortho camera position
-		this.updateCameraClipPlanes( camera );
+		this.adjustCamera( camera );
 
 	}
 
-	// Updates the passed camera near and far clip planes to encapsulate the
-	// ellipsoid from their current position.
-	updateCameraClipPlanes( camera ) {
+	// Updates the passed camera near and far clip planes to encapsulate the ellipsoid from the
+	// current position in addition to adjusting the height.
+	adjustCamera( camera ) {
 
-		const {
-			tilesGroup,
-			ellipsoid,
-		} = this;
+		super.adjustCamera( camera );
 
+		const { tilesGroup, ellipsoid } = this;
 		if ( camera.isPerspectiveCamera ) {
 
+			// adjust the clip planes
 			const distanceToCenter = _vec
 				.setFromMatrixPosition( tilesGroup.matrixWorld )
 				.sub( camera.position ).length();
@@ -270,17 +241,12 @@ export class GlobeControls extends EnvironmentControls {
 			camera.far = distanceToCenter + 0.1;
 
 			// adjust the position of the ortho camera such that the near value is 0
-			if ( ! this.allowNegativeNearPlanes && camera.near < 0 ) {
-
-				camera.position.addScaledVector( _forward, camera.near );
-				camera.far -= camera.near;
-				camera.near = 0;
-
-			}
+			camera.position.addScaledVector( _forward, camera.near );
+			camera.far -= camera.near;
+			camera.near = 0;
 
 			camera.updateProjectionMatrix();
 			camera.updateMatrixWorld();
-
 
 		}
 
@@ -301,7 +267,6 @@ export class GlobeControls extends EnvironmentControls {
 
 			const {
 				enableDamping,
-				inertiaDragMode,
 				tilesGroup,
 				dragQuaternion,
 				dragInertia,
@@ -311,43 +276,39 @@ export class GlobeControls extends EnvironmentControls {
 			// apply inertia for movement
 			if ( enableDamping ) {
 
-				// drag mode 1 means we're near the globe
-				if ( inertiaDragMode === 1 ) {
+				// ensure our w component is non-one if the xyz values are
+				// non zero to ensure we can animate
+				if (
+					dragQuaternion.w === 1 && (
+						dragQuaternion.x !== 0 ||
+						dragQuaternion.y !== 0 ||
+						dragQuaternion.z !== 0
+					)
+				) {
 
-					// ensure our w component is non-one if the xyz values are
-					// non zero to ensure we can animate
-					if (
-						dragQuaternion.w === 1 && (
-							dragQuaternion.x !== 0 ||
-							dragQuaternion.y !== 0 ||
-							dragQuaternion.z !== 0
-						)
-					) {
-
-						dragQuaternion.w = Math.min( dragQuaternion.w, 1 - 1e-9 );
-
-					}
-
-					// construct the rotation matrix
-					_center.setFromMatrixPosition( tilesGroup.matrixWorld );
-					_quaternion.identity().slerp( dragQuaternion, dragInertia.x * deltaTime );
-					makeRotateAroundPoint( _center, _quaternion, _rotMatrix );
-
-					// apply the rotation
-					camera.matrixWorld.premultiply( _rotMatrix );
-					camera.matrixWorld.decompose( camera.position, camera.quaternion, _vec );
-
-				} else if ( inertiaDragMode === - 1 ) {
-
-					this._applyZoomedOutRotation( dragInertia.x * deltaTime, dragInertia.y * deltaTime );
+					dragQuaternion.w = Math.min( dragQuaternion.w, 1 - 1e-9 );
 
 				}
 
+				// construct the rotation matrix
+				_center.setFromMatrixPosition( tilesGroup.matrixWorld );
+				_quaternion.identity().slerp( dragQuaternion, dragInertia.x * deltaTime );
+				makeRotateAroundPoint( _center, _quaternion, _rotMatrix );
+
+				// apply the rotation
+				camera.matrixWorld.premultiply( _rotMatrix );
+				camera.matrixWorld.decompose( camera.position, camera.quaternion, _vec );
+
 			}
 
-		} else if ( this._dragMode === 1 || this._isNearControls() ) {
+		} else {
 
-			this._dragMode = 1;
+			// save the drag mode state so we can update the pivot mesh visuals in "update"
+			if ( this._dragMode === 0 ) {
+
+				this._dragMode = this._isNearControls() ? 1 : - 1;
+
+			}
 
 			const {
 				raycaster,
@@ -405,9 +366,6 @@ export class GlobeControls extends EnvironmentControls {
 			camera.matrixWorld.premultiply( _rotMatrix );
 			camera.matrixWorld.decompose( camera.position, camera.quaternion, _vec );
 
-			// track inertia variables
-			this.inertiaDragMode = 1;
-
 			const { dragInertia, dragQuaternion } = this;
 			if ( pointerTracker.getMoveDistance() / deltaTime < 2 * window.devicePixelRatio ) {
 
@@ -421,130 +379,9 @@ export class GlobeControls extends EnvironmentControls {
 
 			}
 
-		} else {
-
-			this._dragMode = - 1;
-
-			const {
-				pointerTracker,
-				rotationSpeed,
-				camera,
-				pivotMesh,
-				tilesGroup,
-				ellipsoid,
-				domElement,
-			} = this;
-
-			// get the delta movement with magic numbers scaled by the distance to the
-			// grabbed point so it feels okay
-			// TODO: it would be better to properly calculate angle based on drag distance
-			let scaleAmount;
-			if ( camera.isPerspectiveCamera ) {
-
-				// get pointer positions
-				pointerTracker.getCenterPoint( _pointer );
-				pointerTracker.getPreviousCenterPoint( _prevPointer );
-
-				// convert to [0, 1] coords
-				mouseToCoords( _pointer.x, _pointer.y, domElement, _pointer );
-				mouseToCoords( _prevPointer.x, _prevPointer.y, domElement, _prevPointer );
-
-				// project to near plane and take delta
-				_vec.set( _pointer.x, _pointer.y, - 1 ).unproject( camera );
-				_pos.set( _prevPointer.x, _prevPointer.y, - 1 ).unproject( camera );
-				_vec.sub( _pos );
-
-				// find the drag vector at the distance of the ellipsoid
-				const radius = Math.max( ...ellipsoid.radius );
-				const distToSurface = this.getDistanceToCenter() - radius;
-				const scaledDragDist = _vec.distanceTo( _pos ) * distToSurface / camera.near;
-
-				// scale the rotation amount by the radius of the ellipsoid
-				scaleAmount = 7.5 * 1e-4 * scaledDragDist / radius;
-
-			} else {
-
-				scaleAmount = MathUtils.mapLinear(
-					camera.zoom,
-					this._getOrthographicTransitionZoom(),
-					this._getMinOrthographicZoom(),
-					0.001,
-					0.005,
-				);
-
-			}
-
-			pointerTracker.getCenterPoint( _pointer );
-			pointerTracker.getPreviousCenterPoint( _prevPointer );
-			_deltaPointer
-				.subVectors( _pointer, _prevPointer )
-				.multiplyScalar( scaleAmount );
-
-			const azimuth = - _deltaPointer.x * rotationSpeed;
-			const altitude = - _deltaPointer.y * rotationSpeed;
-
-			_center.setFromMatrixPosition( tilesGroup.matrixWorld );
-			_right.set( 1, 0, 0 ).transformDirection( camera.matrixWorld );
-			_up.set( 0, 1, 0 ).transformDirection( camera.matrixWorld );
-
-			// apply the altitude and azimuth adjustment
-			_quaternion.setFromAxisAngle( _right, altitude );
-			camera.quaternion.premultiply( _quaternion );
-			makeRotateAroundPoint( _center, _quaternion, _rotMatrix );
-			camera.matrixWorld.premultiply( _rotMatrix );
-
-			_quaternion.setFromAxisAngle( _up, azimuth );
-			camera.quaternion.premultiply( _quaternion );
-			makeRotateAroundPoint( _center, _quaternion, _rotMatrix );
-			camera.matrixWorld.premultiply( _rotMatrix );
-
-			camera.matrixWorld.decompose( camera.position, camera.quaternion, _vec );
-
-			pivotMesh.visible = false;
-
-			// update drag variables
-			this.inertiaDragMode = - 1;
-			_deltaPointer.multiplyScalar( 1 / deltaTime );
-
-			if ( pointerTracker.getMoveDistance() / deltaTime < 2 * window.devicePixelRatio ) {
-
-				this.dragInertia.lerp( _deltaPointer, 0.5 );
-
-			} else {
-
-				this.dragInertia.copy( _deltaPointer );
-
-			}
-
 		}
 
 		this._alignCameraUp( this.up );
-
-	}
-
-	_applyZoomedOutRotation( x, y ) {
-
-		const { rotationSpeed, tilesGroup, camera } = this;
-		const azimuth = - x * rotationSpeed;
-		const altitude = - y * rotationSpeed;
-		camera.updateMatrixWorld();
-
-		_center.setFromMatrixPosition( tilesGroup.matrixWorld );
-		_right.set( 1, 0, 0 ).transformDirection( camera.matrixWorld );
-		_up.set( 0, 1, 0 ).transformDirection( camera.matrixWorld );
-
-		// apply the altitude and azimuth adjustment
-		_quaternion.setFromAxisAngle( _right, altitude );
-		camera.quaternion.premultiply( _quaternion );
-		makeRotateAroundPoint( _center, _quaternion, _rotMatrix );
-		camera.matrixWorld.premultiply( _rotMatrix );
-
-		_quaternion.setFromAxisAngle( _up, azimuth );
-		camera.quaternion.premultiply( _quaternion );
-		makeRotateAroundPoint( _center, _quaternion, _rotMatrix );
-		camera.matrixWorld.premultiply( _rotMatrix );
-
-		camera.matrixWorld.decompose( camera.position, camera.quaternion, _vec );
 
 	}
 
@@ -623,10 +460,11 @@ export class GlobeControls extends EnvironmentControls {
 			// the zoom speeds are comparable
 			const dist = this.getDistanceToCenter() - ellipsoid.radius.x;
 			const scale = zoomDelta * dist * zoomSpeed * 0.0025;
+			const clampedScale = Math.max( scale, Math.min( this.getDistanceToCenter() - maxDistance, 0 ) );
 
 			// zoom out directly from the globe center
 			this.getVectorToCenter( _vec ).normalize();
-			this.camera.position.addScaledVector( _vec, scale );
+			this.camera.position.addScaledVector( _vec, clampedScale );
 			this.camera.updateMatrixWorld();
 
 			this.zoomDelta = 0;
@@ -643,7 +481,10 @@ export class GlobeControls extends EnvironmentControls {
 			const normalizedDelta = Math.pow( 0.95, Math.abs( scale * 0.05 ) );
 			const scaleFactor = scale > 0 ? 1 / Math.abs( normalizedDelta ) : normalizedDelta;
 
-			camera.zoom = Math.max( this._getMinOrthographicZoom(), Math.min( maxZoom, camera.zoom * scaleFactor * zoomSpeed ) );
+			const maxScaleFactor = minZoom / camera.zoom;
+			const clampedScaleFactor = Math.max( scaleFactor * zoomSpeed, Math.min( maxScaleFactor, 1 ) );
+
+			camera.zoom = Math.min( maxZoom, camera.zoom * clampedScaleFactor );
 			camera.updateProjectionMatrix();
 
 			this.zoomDelta = 0;
@@ -782,7 +623,7 @@ export class GlobeControls extends EnvironmentControls {
 		const orthoSize = Math.min( orthoHeight, orthoWidth );
 		const ellipsoidRadius = Math.max( ...ellipsoid.radius );
 		const ellipsoidDiameter = 2 * ellipsoidRadius;
-		return 0.5 * orthoSize / ellipsoidDiameter;
+		return 0.7 * orthoSize / ellipsoidDiameter;
 
 	}